ch: add poker_room logic
 - add poker_room logic

[Finishes #172825140]
